Remove noise both vertical and horizontal
We have spacer lines after task banners that we don't need.

We don't have spacer lines between a PLAY RECAP and the next PLAY which
we do need.

We have a spacer line before a PLAY RECAP which doubles with the
post-task spacer.

We don't need args printed in task banner lines.

Change-Id: I678690bbb4fc41d6028b5e86f43574989c24f8b6