Fix change history cycle detection
We were short-circuiting loading some changes which appeared in
our dependent change query history because we were only looking
at change numbers, not change+patchset.  Record both in the history
to avoid this problem.

Note, this could likely be simplified further, but rather than doing
so now, I'll leave that for when we implement cross-source dependencies,
which is very likely to change this further.

This also corrects a defect in the fake gerrit change test fixture.

Change-Id: Ie1d9fe0dfd32032caf4ff8c90bded30d5e7bf886