Log items in loops better
If the loop item has a result msg - we should display it.

If the loop item doens't have a result msg - there is no other
identifying feature than the loop variable - which may be a full dict or
other json-object. In that case, pretty print it rather than one-lining
it. This may not be perfect - but let's give it a try.

Change-Id: Ib9a4dc0ac7783559151f12996144941e534633ee