Add zuul.items to job vars
In v2 there was a way for a job to see all the changes being tested
with it.  This is more complicated in v3 with the addition of
multiple sources.  Fortunately, we can pass all the information we
need to accurately represent the situation to the job via an
data structure in ansible variables.

Change-Id: Ibca2065e9014ec67167057f89ef07d36103f89c7