Add spacer line after loops
In order to do this we need to make sure to register loop results. We
also need to register them when things are skipped. And, if the loop
items have been skipped, there will be no status in the deferred result,
so we should not print an All Items Complete message.

Change-Id: I620318df7bb13a8f3d325b0a01296586aca55f1f