Docs: add a glossary
And start using it.  Added some common terms to glossary.

I have not found a lot of source material yet to suggest the best
way to incorporate use of a glossary into the text, especially with
very frequently used technical terms like we have in some parts of
this documentation.  So here are the guidelines I'm trying out:

* If the term is being defined in the text, don't like to the glossary
  (that would be redundant), but do emphasize it with *italics* the
  first time it appears in that definition.  Subsequent uses within
  the same subsection should be in regular type.
* If it's being used (but not defined) in the text, link the first usage
  within a subsection to the glossary, but subsequent uses should be
  in regular type.
* Be cognizant of how readers may jump to link targets within the text,
  so be liberal in considering that once you cross a link target,
  you may be in a new "subsection" for the above guideline.

This change also alters some use of literals to make them more consistent:

* Filenames and scalar values of attributes (not special keyword values
  like 'independent') should be ``literal``.

* Most references to attribute names should be via the :attr: role, however,
  for abbreviated local references (ie, to the current or previous attribute)
  use **bold** to maintain the same style but without the long name and link.

Also, correct some spelling.

Change-Id: I7debd962c0300f1a7c357d825fc2642ada89df58