Limit github reporting to github sources
Also limit it to the specific github host that generated the event. This
corrects an issue seen live where the github driver attempts to report
on gerrit events causing a failure, which will eject the change without
reporting it on the gerrit side.

Testing this required adding start reporters, so that a traceback in
reporting a start to the wrong driver would cause the change to get
dequeued before the jobs run, and before the success is reported.

Also a new test needed to be added to make sure multiple connections
using the github driver are able to correctly report to the right github
server.

Change-Id: Ic4630dd46d72fbeb45f691615a99ef1299a8cd05