Remove getProjectBranches from FakeGithubConnection
The method getProjectBranches from the GithubConnection is currently
not used by any test as it is overridden in
FakeGithubConnection. Remove this from here and inject a FakeGithub
instead so this method gets tested. This could be improved in the
future to also remove more overrides and test more of GithubConnection
through the FakeGithub.

Change-Id: I0135cd0cc04857abcb93ce73dd5b976a27d26e1d