Add zuul:var directive and role
This adds a directive (and role) for zuul variables, which also
have a complex hierarchy like configuration values.

The type field may be used to cause, for example, 'items' to render
as 'items[]' to remind users that they must provide an index.  It
is ignored for the purposes of the cross-reference role, indices,
and link targets.

Change-Id: I674b17d8257c501cfc5c2f29f2aab9804e9ec846