Docs: change default title for value xref
When we use zuul:value to xref, the default title text for the link
is the full target name (eg "pipeline.trigger.github.event.pull_request")
which is a bit verbose.  Normally, we would just want the final element
as the title for a value (so a sentence might read "When a <pull_request>
event is received...").

Change-Id: I42cd678d9bb97058d5fbe255791cbc476d5057bb