Update config cache only after all cat jobs complete
If we encounter an error when reconfiguring, we generally roll back
to the prior configuration.  However, within one of the load functions,
we clear the cached data on each project, then send out all the cat
jobs, then update the cached data with the results.  If an error occurs
between clearing and updating the cache, we will be left with some
(but not all) projects missing their cache.  This will prevent future
dynamic reconfigurations from completing.

To remedy this, don't update the cache until the end of the load
function to ensure that all of the tricky network stuff has
completed successfully.

Change-Id: I3bf60a93e57105418037b70156605a8a514dab24