Added more debug logging for github requests
Add two attributes to the github object storing user and project for the
rate limit lines. Also add a log line before every rate limit line that
explains what we just did. This could likely be improved by logging
BEFORE we do a thing then letting the rate limit line serve as
confirmation that we did that thing, but I didn't think of that until
writing this commit message.

Change-Id: Ib7bc9392a5e4bf4dd1bef294192fd5dc3107797c