Add helper script for doing local log streaming tests
This isn't 100% of a local run of the functional test- but it's
basically how I've been testing stuff locally and runs the playbooks
that the functional test runs.

It takes an optional parameter to an inventory that has a controller
node, 2 nodes named node1 and node2 and a group called node that
contains node1 and node2. If one isn't provided, it generates one with
all of those hosts pointing to localhost.

Log output will go into test-logs-output. Subsequent runs will append to
the log file, so one can keep a tail -f in another window.

Change-Id: I46afd26a6fdaa348e9753fe0f3d3c1b194766df4