Fix dynamic dependent pipeline failure
When a change that altered a dependent pipeline failed and was removed
from that pipeline, changes behind it did not have their configuration
reset; instead, the alteration introduced by the failed change lingered.
This was largely because the way we tested that a project participated
in a pipeline was incorrect -- we only checked that it was in the layout,
not the pipeline.  That is corrected.

Additionally, the job graph for the second change was not reset.  That's
because it was only created once, and was not cleared out when the
change's buildset is reset.  That is corrected as well (though a future
enhancement would be to move item.job_graph into the BuildSet class to
better reflect its lifecycle).

Change-Id: Icdbc17bb21887a96f4118bb1a49ed7b00e1304e6